add more game logic

+++ Rlatro.Core/Contracts/Consumables/IUsageCondition.cs
@@ -4,6 +4,6 @@ namespace Balatro.Core.Contracts.Consumables
 {
     public interface IUsageCondition
     {
-        bool IsUsable(GameContext context, byte[] cardIndexes);
+        bool IsUsable(GameContext context, int[] cardIndexes);
     }
 }
